Think it should be pretty much all good now!
So I changed over to using the iRacingSdkWrapper instead of dealing with sdk directly. Not sure if that directly fixed anything, but due to some helpful debug outputs it  indirectly led me to find why it seems to have trouble connecting sometimes! I was apparantly using an outdated version of the C# sdk implementation. Not sure if that's what was causing the problem before or not, but once I updated it after having switched everything over to the wrapper implementation it seems to be working without a hitch!

The Wrapper should also make adding any more features in the future a lot simpler! One of the best things it does compared to what I was doing before is it executes my code when the sdk sends out it's telemetry update instead of my check every 5 ms to try and keep it as up to date as possible.