Save program settings added
Adde dthe ability to save settings. The config file is saved in %userprofile%\appdata\local\Edward_Frederick-Bittner (And then another sub folder). Not sure why but I needed to add a OverlayController.exe.config with all empty values to get it to start... But the values that the programs reads from are in the appdata folder.